requirements.txt
nibabel
numpy
tqdm
scipy
matplotlib
SimpleITK
torchio
antspyx
streamlit